Actualización de scripts SQL y creación de scripts de popular y extraer

+++ BaseDeDatos/data/crear_tabla_ca.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS comunidades_autonomas (
-    codigo_ine SMALLINT UNSIGNED PRIMARY KEY,
+    codigo_ine TINYINT UNSIGNED PRIMARY KEY,
     nombre VARCHAR(50) NOT NULL  
 );
